fix init column bug
