+++ fe/src/test/java/org/apache/doris/analysis/DataDescriptionTest.java
@@ -76,7 +76,7 @@ public class DataDescriptionTest {
         desc.analyze("testDb");
         Assert.assertEquals("DATA INFILE ('abc.txt') NEGATIVE INTO TABLE testTable (col1, col2)", desc.toString());
         Assert.assertEquals("testTable", desc.getTableName());
-        Assert.assertEquals("[col1, col2]", desc.getColumnNames().toString());
+        Assert.assertEquals("[col1, col2]", desc.getFileFieldNames().toString());
         Assert.assertEquals("[abc.txt]", desc.getFilePaths().toString());
         Assert.assertTrue(desc.isNegative());
         Assert.assertNull(desc.getColumnSeparator());
